read in serial, refactor
