+++ pkg/disttask/importinto/BUILD.bazel
@@ -63,7 +63,6 @@ go_library(
         "//pkg/util/disttask",
         "//pkg/util/etcd",
         "//pkg/util/logutil",
-        "//pkg/util/mathutil",
         "//pkg/util/promutil",
         "//pkg/util/redact",
         "@com_github_docker_go_units//:go-units",
@@ -72,7 +71,6 @@ go_library(
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_prometheus_client_golang//prometheus",
-        "@com_github_tikv_client_go_v2//error",
         "@com_github_tikv_client_go_v2//oracle",
         "@com_github_tikv_client_go_v2//util",
         "@org_golang_x_exp//maps",
